ENH Update models
Use fixed training sets: no repeated sequences

Use scikit-learn

Use changed feature set

Fix testing: The training and testing overlapped significantly!
Therefore, we need to be very careful in testing. Thus, we replace the
test results for the overlapping sequences with the out-of-bag estimates
for those same sequences.

As the results changed, the tests needed to change as well