ENH Better information for installation  (#66)
* ENH fix pyrodigal version avoid error

Newer versions of pyrodigal uses different command names for the old functions in the script, keeping the results. So, we fix the version to avoid breaking the code with different python versions.

* ENH Include specification for dedicated environment

Include instructions for installation of macrel inside a dedicated environment as the preferred method to work with Macrel. That is important to avoid conflict of common packages that most of the programs use in different versions that may conflict inside the Macrel's structure.

* ENH installation instructions clarity

make the recipe for installation using conda in a dedicated environment more clear. ---